refactor: make formatter loading asynchronous
Refactor to asynchronous formatter loading because the new ESLint class
loads formatters asynchronously.
Move formatter loading to the linter file for access to the ESLint instance
because the new ESLint loadFormatter method is not static.
Refactor formatter loading to use a CLIEngine instance.
Refactor the linting function to load the formatters.

Preperation for migration to the ESLint class, see #10.