refactor: add loader option filtering
Refactor to filter out loader options before passing the options to ESLint
because the new ESLint class throws when unknown options are passed to it.
Add method for filtering out loader options and unit tests for it.
Rename the getOptions file options because it now contains option utilities.

It was possible to have getESLintOptions as a private function in
getCLIEngine, but having it exported helps testing.
There is no way to query the ESLint class for the options it received so
there is no way to test it through the getCLIEngine/ESLint method.

Part of the migration to the new ESLint class, see #10.