refactor: add custom formatter mock
Refactor the custom formatter tests to use a mock.
Refactor to test the error output to make sure the selected formatter
is being used.
The test used to always pass because formatter loading fails silently and
returns the default formatter.
Add a mock formatter.
Remove eslint-friendly-formatter that was only used for the test.

This refactor helps decouple the tests from the results format.
Making the migration to the new ESLint class easier.

Preperation for migration to the ESLint class, see #10.