refactor: migrate tests to the ESLint class
Refactor the ESLint mock to imitate the new ESLint class.
Refactor to simplify mock formatter loading.
Refactor ESLint options because the new API introduces a new options format.

The tests do not pass yet as the plugin has not been refactored.

Part of the migration to the new ESLint class, see #10.