refactor: migrate to the new ESLint class
Refactor linter to the new ESLint class.
Refactor the getCLIEngine method to getESLint.
Refactor LintDirtyModulesPlugin to remove use of resolveFileGlobPatterns
which is not available in the new API. File paths should work fine as
globs and do not need to be converted.

Completing the migration to the new ESLint class, see #10.