refactor: simplify dirty file linting
Refactor LintDirtyModulesPlugin to only handle dirty file filtering.
The refactor allows for easier testing without the need for mocks.
Rename LintDirtyModulesPlugin to DirtyFileWatcher to make the name more
descriptive of the new functionality.
Refactor the tests for the simplified DirtyFileWatcher.
Refactor the tests to use absolute paths because Webpack passes absolute paths
in the compiler.fileTimestamps map.
Refactor to only test system specific paths. The functionality of the requested
change requires using real paths for some tests which causes issues with
testing both Windows and Unix paths on the same machine.
As the CI runs on Windows, Mac and Linux paths will be tested on all platforms.

Preparation for requested change in #11.