perf: parse patterns to globs only on startup
Refactor to parse patterns only on startup instead of parsing them to
globs on each run.
In watch mode Webpack does not watch it's own configuration so the
plugin configuration does not change when Webpack is running.
Because the file patterns and extensions do not change there is no need
to process them into globs separately for each run.